backend/requirements.txt
# BCK Manager Web — Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-crontab>=3.0.0
asyncssh>=2.14.0
httpx>=0.25.0
packaging>=23.0
PyYAML>=6.0
boto3>=1.28.0
aiofiles>=23.0
